requirements.txt

pyserial

# for fw-pack.py
crcmod
